requirements.txt
# Weather Dashboard Dependencies

# Core libraries
requests>=2.31.0
matplotlib>=3.7.0
seaborn>=0.12.0
flet>=0.10.0

# Environment variable management
python-dotenv>=1.0.0

# Optional: For better error handling and logging
urllib3>=2.0.0
